melt test is in own module
